requirements.txt
openai
python-dotenv
groq
pandas